+++ tango/__init__.py
@@ -27,7 +27,7 @@ from .shapes import (
     TypeGenerator,
 )
 
-__version__ = "0.4.0"
+__version__ = "0.4.1"
 __all__ = [
     "TangoClient",
     "TangoAPIError",